backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
groq==0.11.0
httpx==0.27.2
vaderSentiment==3.3.2
pytest==7.4.3
hypothesis==6.92.1
python-dotenv==1.0.0
PyJWT==2.8.0
